% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_arrow.R
\name{define.arrow}
\alias{define.arrow}
\title{Helper function to calculate start and end points of directed edges}
\usage{
define.arrow(x.0, y.0, x.1, y.1, radius = 0.1)
}
\arguments{
\item{x.0}{x coordinate of start point}

\item{y.0}{y coordinate of start point}

\item{x.1}{x coordinate of end point}

\item{y.1}{y coordinate of end point}

\item{radius}{radius around nodes that should not be used for drawing edges}
}
\value{
data.frame
}
\description{
Defines a radius around each node. Edges only start and end outside of
this radius
}
\examples{
define.arrow(
  x.0 = -.5, 
  y.0 = 0, 
  x.1 = .5, 
  y.1 = 0, 
  radius = .1
)
}
