% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edges}
\alias{edges}
\title{Edges of an example DAG.}
\format{
A data frame with variables: \code{from}, \code{to},
\code{lwd} and \code{col}.
}
\usage{
edges
}
\description{
Edges of an example DAG.
}
\keyword{datasets}
