% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Nodes of an example DAG.}
\format{
A data frame with variables: \code{name}, \code{description},
\code{x} and \code{y}.
}
\usage{
nodes
}
\description{
Nodes of an example DAG.
}
\keyword{datasets}
